Closes #8, closes #12
Images and GIFs are prefixed with a timestamp
All the conversations are now retrieved (not only the 20 or 50 latest). The number of found threads is displayed.
The path of the written conversation is displayed in the console
Emojis are now written as Unicode instead of their description
Dedicated behavior for macOS (Darwin) for emojis is removed
Description (alt text) of stickers is written in the log
Update of all dependencies
Binaries are built with Python 3.4 (downgrade from 3.5) to improve compatibility
Version and Python version and platform are displayed at startup
Refactoring (pep8, lint)