Initial API-centric/JSON overhaul, 2FA login.
With #83, we need a new approach to allow this program to function.
This is the first attempt. Some features have been broken or removed,
and likely cannot be added back. Cards, embedded tweets, etc. have been
dropped, and it's possible that stickers are broken too (do they even
exist anymore?).

I can't promise that this works robustly, but it was tested in Python
3.7 with a saved session; I'm also not sure authentication isn't
totally broken as I tried to implement a 2FA fix, too, but locked
my account out from too many login attempts before I managed to get it
working.